+++ admin/src/components/CategoryList/index.tsx
@@ -1,13 +1,17 @@
 import React, { useState, useEffect } from 'react';
 
-import { useFetchClient, useNotification } from '@strapi/helper-plugin';
+import { useFetchClient, useNotification, useCMEditViewDataManager } from '@strapi/helper-plugin';
 
 import { SingleSelect, SingleSelectOption } from '@strapi/design-system';
 
+import { useIntl, MessageDescriptor } from 'react-intl';
+
 type CategoryListProps = {
+  contentTypeUID: string;
   onChange: (event: { target: { name: string; value?: string | null; type?: string } }) => void;
   name: string;
-  description?: string;
+  intlLabel: MessageDescriptor;
+  hint?: string;
   disabled?: boolean;
   error?: string;
   required?: boolean;
@@ -20,15 +24,19 @@ type CategoryModelType = {
 };
 
 export function CategoryList({
+  contentTypeUID,
   name,
+  intlLabel,
   onChange,
   value,
   required,
   error,
-  description,
+  hint,
   disabled,
 }: CategoryListProps) {
-  const { get } = useFetchClient();
+  const { get, post } = useFetchClient();
+  const { initialData } = useCMEditViewDataManager();
+  const { formatMessage } = useIntl();
 
   const toggleNotification = useNotification();
 
@@ -73,12 +81,38 @@ export function CategoryList({
     fetchData();
   }, []);
 
+  useEffect(() => {
+    async function validateSlug(productSlug: string) {
+      try {
+        const { data } = await post(`/commercetools/url/validate`, {
+          contentTypeUID,
+          slug: productSlug,
+          initialSlug: initialData?.[name],
+          field: name,
+        });
+
+        if (data.isValid) {
+          setErrorMessage(undefined);
+        } else {
+          setErrorMessage('This category page already exists.');
+        }
+      } catch (e) {
+        console.error(e);
+      }
+    }
+
+    if (value) {
+      validateSlug(value);
+    }
+  }, [value]);
+
   return (
     <SingleSelect
-      label={name}
+      label={formatMessage(intlLabel)}
+      name={name}
       required={required}
       placeholder={isLoading ? 'Loading...' : 'Select a category'}
-      hint={description}
+      hint={hint}
       error={errorMessage}
       disabled={disabled || isLoading}
       value={value}