+++ server/controllers/index.ts
@@ -1,7 +1,9 @@
 import productController from './product-controller';
 import categoryController from './category-controller';
+import urlController from './url-controller';
 
 export default {
   productController,
   categoryController,
+  urlController,
 };