+++ server/routes/index.ts
@@ -17,8 +17,33 @@ export default [
   },
   {
     method: 'GET',
+    path: '/getProductBySku/:id',
+    handler: 'productController.getProductBySku',
+    config: {
+      policies: [],
+    },
+  },
+  {
+    method: 'GET',
     path: '/getAllCategories',
     handler: 'categoryController.getAllCategories',
+    config: {
+      policies: [],
+    },
+  },
+  {
+    method: 'POST',
+    path: '/url/generate',
+    handler: 'urlController.getUrlBySlug',
+    config: {
+      policies: [],
+    },
+  },
+
+  {
+    method: 'POST',
+    path: '/url/validate',
+    handler: 'urlController.validateSlug',
     config: {
       policies: [],
     },