+++ CHANGELOG.md
@@ -1,3 +1,104 @@
+# [1.18.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.17.0...v1.18.0) (2024-04-16)
+
+
+### Features
+
+* Add about us page URL to URL service ([fea1c9c](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/fea1c9c29e707ffb96b41b850fa9c0c0d0d8460f))
+
+# [1.17.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.16.0...v1.17.0) (2024-04-10)
+
+
+### Features
+
+* Refactor category filtering and ordering ([cda34e6](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/cda34e68b69ca8b48ea4bed3b596075f92726d74))
+
+# [1.16.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.15.1...v1.16.0) (2024-03-20)
+
+
+### Features
+
+* Add slug validation to CategoryList and ProductGrid components ([1a4c52e](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/1a4c52ee6facd67e755bb73c7419656ef586e4e8))
+* Add URL slug validation endpoint ([53f9f9b](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/53f9f9bd6454f1d31b28ef63bc919232cc89e6e2))
+
+## [1.15.1](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.15.0...v1.15.1) (2024-03-20)
+
+
+### Bug Fixes
+
+* Update initial selected product state ([882ecfd](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/882ecfd069c58e2590a2e7fe99c17a95ba757591))
+
+# [1.15.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.14.0...v1.15.0) (2024-03-15)
+
+
+### Bug Fixes
+
+* Revert image sort on products ([15d6492](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/15d6492c9abcb824aa830e002b42477bc8aa8ec5))
+* Update pagination data ([1871d74](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/1871d74e2697ce0da6338805f39d00b94b0bb289))
+
+
+### Features
+
+* Add react-intersection-observer dependency ([b578a98](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/b578a9897b92ef95b6172bdb51e66cc182d8649f))
+* Add search functionality to ProductGridModal component ([d3ccc74](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/d3ccc74d4c93af72ca2ed8830dea577cf82a61ca))
+* Add use-debounce dependency ([9901b3b](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/9901b3b16d2e1fb40d726bc75f6144293bf480d4))
+* Refactor product grid modal and pagination ([9e86071](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/9e86071373022e95f54aacd5777dca5c72eb8572))
+
+# [1.14.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.13.1...v1.14.0) (2024-03-14)
+
+
+### Bug Fixes
+
+* Add label to category list ([cece5f9](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/cece5f9bfa233c671caf322a834cdbe35d6f5d75))
+
+
+### Features
+
+* Add ancestors slugs to url service ([ba9aa6c](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/ba9aa6c3cfa608983a67cc51a19d495148c9ac8e))
+* Refactor getAllCategories method to limit to 100 ([fd21fff](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/fd21fff583c5974e88deac7fa98f43583ef80079))
+
+## [1.13.1](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.13.0...v1.13.1) (2024-03-14)
+
+
+### Bug Fixes
+
+* Remove console logs ([cb8c6d6](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/cb8c6d6654352f84329520d8eb8bce408e4216e5))
+
+# [1.13.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.12.0...v1.13.0) (2024-03-14)
+
+
+### Features
+
+* Add admin for the UrlTextField ([d84f2f1](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/d84f2f1f6f813a3e4c390d2c1cd7670950566fbc))
+* Add url service/controller/route ([9cd0822](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/9cd0822386ce20deecf24580beaf8e0e12c1f21b))
+
+# [1.12.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.11.0...v1.12.0) (2024-02-28)
+
+
+### Features
+
+* Order categories by name in category controller ([d30dd8c](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/d30dd8c0f3f8b4451a8601413446ba21b9b6bf0b))
+
+# [1.11.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.10.0...v1.11.0) (2024-02-28)
+
+
+### Features
+
+* Refactor product controller to sort products by image count. Add types ([f3e83a9](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/f3e83a9ebae002c9cb4afdfc771e832546c6e370))
+
+# [1.10.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.9.0...v1.10.0) (2024-02-19)
+
+
+### Features
+
+* Refactor category controller to include parent categories in the name ([097d2cc](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/097d2cc08a9da274970906d29c72a425932dac95))
+
+# [1.9.0](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.8.2...v1.9.0) (2024-01-31)
+
+
+### Features
+
+* Add price placeholder to admin ([f37ca0b](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/commit/f37ca0b58dd67f46f06f9badfed6286f69c98480))
+
 ## [1.8.2](https://github.com/JonssonWorkwear/strapi-plugin-commercetools-integration/compare/v1.8.1...v1.8.2) (2024-01-31)